requirements.txt
Flask
pymongo
python-dotenv
gunicorn
```